python/examples/langchain/requirements.txt
lightspark
langchain
langchain-openai
langgraph
python-dotenv
